Celestial Bodies

AstroBase.Bodies.Barycenter — Type

Barycenter <: CelestialBody

Abstract supertype for representing the barycenters of the solar system and planetary systems as pseudo-bodies.

AstroBase.Bodies.CelestialBody — Type

CelestialBody

Abstract supertype for all celestial bodies and pseudo-bodies.

AstroBase.Bodies.MinorBody — Type

MinorBody <: CelestialBody

Abstract supertype for minor solar system bodies.

AstroBase.Bodies.NAIFId — Type

NAIFId

An integer code for identifying celestial bodies and other objects in space.

References

NASA NAIF

AstroBase.Bodies.NaturalSatellite — Type

NaturalSatellite <: CelestialBody

Abstract supertype for natural satellites (moons).

AstroBase.Bodies.Planet — Type

Planet <: CelestialBody

Abstract supertype for planets.

AstroBase.Bodies.along_orbit_radius — Function

mean_radius([T=Float64,] body::CelestialBody)

Return the mean radius for body in km.

Example

julia> along_orbit_radius(earth)
6378.1366

AstroBase.Bodies.equatorial_radius — Function

mean_radius([T=Float64,] body::CelestialBody)

Return the mean radius for body in km.

Example

AstroBase.Bodies.grav_param — Function

grav_param([T=Float64,] body::CelestialBody)

Return the gravitational parameter $\mu = GM$ for body in km^3/s^2.

Example

julia> grav_param(earth)
398600.435436096

AstroBase.Bodies.mean_radius — Function

mean_radius([T=Float64,] body::CelestialBody)

Return the mean radius for body in km.

Example

julia> mean_radius(earth)
6371.008366666666

AstroBase.Bodies.polar_radius — Function

mean_radius([T=Float64,] body::CelestialBody)

Return the mean radius for body in km.

Example

AstroBase.Bodies.subplanetary_radius — Function

mean_radius([T=Float64,] body::CelestialBody)

Return the mean radius for body in km.

Example