Constants

AstroBase.Constants.astronomical_unit — Method

astronomical_unit([[T=Float64], unit=km])

Return the astronomical unit $au$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.drift_rate — Method

drift_rate([T=Float64,] TDB, TCB)

Return the rate of drift between TDB and TCB $L_B$.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.drift_rate — Method

drift_rate([T=Float64,] TCG, TCB)

Return the rate of drift between TCG and TCB $L_C$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.drift_rate — Method

drift_rate([T=Float64,] TT, TCG)

Return the rate of drift between TT and TCG $L_G$.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.earth_rotation_angle_j2000 — Method

earth_rotation_angle_j2000([T=Float64])

Return the Earth rotation angle at J2000.0 $\theta_0$.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.earth_rotation_rate — Method

earth_rotation_rate([T=Float64])

Return the rate of advance of Earth rotation angle $d\theta/dUT1$.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.equatorial_radius_earth — Method

equatorial_radius_earth([[T=Float64], unit=km])

Return the equatorial radius of Earth $a\_E$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.gaussian_gravitational_const — Method

gaussian_gravitational_const([T=Float64])

Return the Gaussian gravitational constant $k$.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.geocentric_gravitational_const — Method

geocentric_gravitational_const([[[T=Float64], scale=TDB], unit=km])

Return the geocentric gravitational constant $GM_E$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.geoid_potential — Method

geoid_potential([T=Float64])

Return $J_2$ dynamical form factor of Earth and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.gravitational_const — Method

gravitational_const([[T=Float64], unit=km])

Return the gravitational constant $G$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.heliocentric_gravitational_const — Method

heliocentric_gravitational_const([[[T=Float64], scale=TDB], unit=km])

Return the heliocentric gravitational constant $GM_S$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.j2_factor_earth — Method

j2_factor_earth([T=Float64])

Return $J_2$ dynamical form factor of Earth and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.j2_rate_earth — Method

j2_rate_earth([T=Float64])

Return the time rate of change in $J_2$ of Earth and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.mass_ratio_ceres — Method

mass_ratio_ceres([T=Float64])

Return the ratio of the mass of Ceres to the mass of the Sun $M_{Ceres}/M_S$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.mass_ratio_eris — Method

mass_ratio_eris([T=Float64])

Return the ratio of the mass of the Sun to the mass of Eris $M_S/M_{Eris}$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.mass_ratio_jupiter — Method

mass_ratio_jupiter([T=Float64])

Return the ratio of the mass of the Sun to the mass of Jupiter $M_S/M_{J}$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.mass_ratio_mars — Method

mass_ratio_mars([T=Float64])

Return the ratio of the mass of the Sun to the mass of Mars $M_S/M_{Ma}$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.mass_ratio_mercury — Method

mass_ratio_mercury([T=Float64])

Return the ratio of the mass of the Sun to the mass of Mercury $M_S/M_{Me}$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.mass_ratio_moon — Method

mass_ratio_moon([T=Float64])

Return the ratio of the mass of the Moon to the mass of Earth $M_M/M_E$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.mass_ratio_neptune — Method

mass_ratio_neptune([T=Float64])

Return the ratio of the mass of the Sun to the mass of Neptune $M_S/M_{N}$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.mass_ratio_pallas — Method

mass_ratio_pallas([T=Float64])

Return the ratio of the mass of Pallas to the mass of the Sun $M_{Pallas}/M_S$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.mass_ratio_pluto — Method

mass_ratio_pluto([T=Float64])

Return the ratio of the mass of the Sun to the mass of Pluto $M_S/M_{P}$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.mass_ratio_saturn — Method

mass_ratio_saturn([T=Float64])

Return the ratio of the mass of the Sun to the mass of Saturn $M_S/M_{Sa}$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.mass_ratio_uranus — Method

mass_ratio_uranus([T=Float64])

Return the ratio of the mass of the Sun to the mass of Uranus $M_S/M_{U}$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.mass_ratio_venus — Method

mass_ratio_venus([T=Float64])

Return the ratio of the mass of the Sun to the mass of Venus $M_S/M_{Ve}$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.mass_ratio_vesta — Method

mass_ratio_vesta([T=Float64])

Return the ratio of the mass of Vesta to the mass of the Sun $M_{Vesta}/M_S$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.mean_angular_velocity_earth — Method

mean_angular_velocity_earth([T=Float64])

Return nominal mean angular velocity of Earth $\omega$.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.obliquity_j2000 — Method

obliquity_j2000([T=Float64])

Return the ratio of the mass of the Moon to the mass of Earth $M_M/M_E$ and the associated uncertainty.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.speed_of_light — Method

speed_of_light([[T=Float64], unit=km])

Return the speed of light $c$.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.

AstroBase.Constants.tdb0 — Method

tdb0([T=Float64])

Return TDB–TCB at T0.

Reference

Luzum, Brian, et al. "The IAU 2009 system of astronomical constants: the report of the IAU working group on numerical standards for Fundamental Astronomy." Celestial Mechanics and Dynamical Astronomy 110.4 (2011): 293.