Reference Frames

AstroBase.Frames.cirf — Constant

cirf

The singleton instance of the CIRF type representing the Celestial Intermediate Reference Frame (CIRF).

Reference

SOFA

AstroBase.Frames.icrf — Constant

icrf

The singleton instance of the ICRF type representing the International Celestial Reference Frame (ICRF).

Reference

SOFA

AstroBase.Frames.itrf — Constant

itrf

The singleton instance of the ITRF type representing the International Terrestrial Reference Frame (ITRF).

Reference

SOFA

AstroBase.Frames.tirf — Constant

tirf

The singleton instance of the TIRF type representing the Terrestrial Intermediate Reference Frame (TIRF).

Reference

SOFA

AstroBase.Frames.CIRF — Type

CIRF

A type representing the Celestial Intermediate Reference Frame (CIRF).

Reference

SOFA

AstroBase.Frames.ICRF — Type

ICRF

A type representing the International Celestial Reference Frame (ICRF).

Reference

SOFA

AstroBase.Frames.ITRF — Type

ITRF

A type representing the International Terrestrial Reference Frame (ITRF).

Reference

SOFA

AstroBase.Frames.TIRF — Type

TIRF

A type representing the Terrestrial Intermediate Reference Frame (TIRF).

Reference

SOFA

AstroBase.Frames.@frame — Macro

@frame(name, type=NAMEFrame, parent=nothing, rotating=false)

Define a new reference frame name which is the singleton instance of type. Optionally provide a parent frame and indicate whether the frame is rotating.

Example

julia> @frame inertial

julia> isinertial(inertial)
true

julia> typeof(inertial)
INERTIALFrame

julia> @frame rotating type=Rotating parent=inertial rotating=true

julia> isrotating(rotating)
true

julia> typeof(rotating)
Rotating