Interfaces

AstroBase.Interfaces.AbstractEphemeris — Type

AbstractEphemeris

Abstract supertype for ephemerides.

AstroBase.Interfaces.AbstractFrame — Type

AbstractFrame

Abstract supertype for all reference frames.

AstroBase.Interfaces.Barycenter — Type

Barycenter <: CelestialBody

Abstract supertype for representing the barycenters of the solar system and planetary systems as pseudo-bodies.

AstroBase.Interfaces.CelestialBody — Type

CelestialBody

Abstract supertype for all celestial bodies and pseudo-bodies.

AstroBase.Interfaces.InertialFrame — Type

InertialFrame

Abstract supertype for (pseudo-)inertial reference frames.

AstroBase.Interfaces.MinorBody — Type

MinorBody <: CelestialBody

Abstract supertype for minor solar system bodies.

AstroBase.Interfaces.NAIFId — Type

NAIFId

An integer code for identifying celestial bodies and other objects in space.

AstroBase.Interfaces.NaturalSatellite — Type

NaturalSatellite <: CelestialBody

Abstract supertype for natural satellites (moons).

AstroBase.Interfaces.Planet — Type

Planet <: CelestialBody

Abstract supertype for planets.

AstroBase.Interfaces.RotatingFrame — Type

RotatingFrame

Abstract supertype for rotating reference frames.

AstroBase.Interfaces.grav_param — Function

grav_param([T=Float64,] body::CelestialBody)

Return the gravitational parameter $\mu = GM$ for body in km^3/s^2.

AstroBase.Interfaces.isinertial — Method

isinertial(frame)

Return true if frame is a (pseudo-)inertial reference frame.

AstroBase.Interfaces.isrotating — Method

isrotating(frame)

Return true if frame is a rotating reference frame.