Utilities

AstroBase.Util.days_to_hms — Method

days_to_hms(days)

Split days into hours, minutes, and seconds.

Example

julia> days_to_hms(0.314159)
(7, 32, 23.337600000000265)

AstroBase.Util.deg_to_dms — Method

deg_to_dms(x)

Split x into degrees, arcminutes, and arcseconds.

Example

julia> deg_to_dms(90.314159)
(90, 18, 50.97240000001307)

AstroBase.Util.deg_to_rad — Method

deg_to_rad(x)

Convert x from degrees to radians.

Example

julia> deg_to_rad(90)
1.5707963267948966

AstroBase.Util.deg_to_sec — Method

deg_to_sec(x)

Convert x from degrees to arcseconds.

Example

julia> deg_to_sec(90.0)
324000.0

AstroBase.Util.dms_to_deg — Method

dms_to_deg(d, m, s)

Convert degrees d, arcminutes m, and arcseconds s to degrees.

Example

julia> dms_to_deg(90, 18, 50.97240000001307)
90.314159

AstroBase.Util.dms_to_rad — Method

dms_to_rad(d, m, s)

Convert degrees d, arcminutes m, and arcseconds s to radians.

Example

julia> dms_to_rad(25, 42, 51.42857142857508)
0.4487989505128276

AstroBase.Util.hms_to_days — Method

hms_to_days(h, m, s)

Convert hours h, minutes m, and seconds s to days.

Example

julia> hms_to_days(7, 32, 50.2)
0.3144699074074074

AstroBase.Util.normalize2pi — Function

normalize2pi(angle[, center=0.0])

Normalize angle to be in the interval [center-π, center+π].

AstroBase.Util.rad_to_deg — Method

rad_to_deg(x)

Convert x from radians to degrees.

Example

julia> rad_to_deg(π/2)
90.0

AstroBase.Util.rad_to_dms — Method

rad_to_dms(x)

Convert x from radians into degrees, arcminutes, and arcseconds.

Example

julia> rad_to_dms(π/7)
(25, 42, 51.42857142857508)

AstroBase.Util.rad_to_sec — Method

rad_to_sec(x)

Convert x from radians to arcseconds.

Example

julia> rad_to_sec(π/2)
324000.0

AstroBase.Util.sec_to_deg — Method

sec_to_deg(x)

Convert x from arcseconds to degrees.

Example

julia> sec_to_deg(324000)
90.0

AstroBase.Util.sec_to_rad — Method

sec_to_rad(x)

Convert x from arcseconds to radians.

Example

julia> sec_to_rad(324000)
1.5707963267948966

AstroBase.Util.spherical_to_cartesian — Method

spherical_to_cartesian(theta, phi)

Convert spherical coordinates to Cartesian.

Arguments

theta: longitude angle in radians

phi: latitude angle in radians

Returns

x: magnitude of projection on x axis

y: magnitude of projection on y axis

z: magnitude of projection on z axis

References

ERFA