Celestial Bodies

AstroBase.Bodies.adrastea — Constant

adrastea

The singleton instance of the Adrastea type.

AstroBase.Bodies.aegaeon — Constant

aegaeon

The singleton instance of the Aegaeon type.

AstroBase.Bodies.aegir — Constant

aegir

The singleton instance of the Aegir type.

AstroBase.Bodies.aitne — Constant

aitne

The singleton instance of the Aitne type.

AstroBase.Bodies.albiorix — Constant

albiorix

The singleton instance of the Albiorix type.

AstroBase.Bodies.amalthea — Constant

amalthea

The singleton instance of the Amalthea type.

AstroBase.Bodies.ananke — Constant

ananke

The singleton instance of the Ananke type.

AstroBase.Bodies.anthe — Constant

anthe

The singleton instance of the Anthe type.

AstroBase.Bodies.aoede — Constant

aoede

The singleton instance of the Aoede type.

AstroBase.Bodies.arche — Constant

arche

The singleton instance of the Arche type.

AstroBase.Bodies.ariel — Constant

ariel

The singleton instance of the Ariel type.

AstroBase.Bodies.atlas — Constant

atlas

The singleton instance of the Atlas type.

AstroBase.Bodies.autonoe — Constant

autonoe

The singleton instance of the Autonoe type.

AstroBase.Bodies.bebhionn — Constant

bebhionn

The singleton instance of the Bebhionn type.

AstroBase.Bodies.belinda — Constant

belinda

The singleton instance of the Belinda type.

AstroBase.Bodies.bergelmir — Constant

bergelmir

The singleton instance of the Bergelmir type.

AstroBase.Bodies.bestla — Constant

bestla

The singleton instance of the Bestla type.

AstroBase.Bodies.bianca — Constant

bianca

The singleton instance of the Bianca type.

AstroBase.Bodies.borrelly — Constant

borrelly

The singleton instance of the Borrelly type.

AstroBase.Bodies.caliban — Constant

caliban

The singleton instance of the Caliban type.

AstroBase.Bodies.callirrhoe — Constant

callirrhoe

The singleton instance of the Callirrhoe type.

AstroBase.Bodies.callisto — Constant

callisto

The singleton instance of the Callisto type.

AstroBase.Bodies.calypso — Constant

calypso

The singleton instance of the Calypso type.

AstroBase.Bodies.carme — Constant

carme

The singleton instance of the Carme type.

AstroBase.Bodies.carpo — Constant

carpo

The singleton instance of the Carpo type.

AstroBase.Bodies.ceres — Constant

ceres

The singleton instance of the Ceres type.

AstroBase.Bodies.chaldene — Constant

chaldene

The singleton instance of the Chaldene type.

AstroBase.Bodies.charon — Constant

charon

The singleton instance of the Charon type.

AstroBase.Bodies.cordelia — Constant

cordelia

The singleton instance of the Cordelia type.

AstroBase.Bodies.cressida — Constant

cressida

The singleton instance of the Cressida type.

AstroBase.Bodies.cupid — Constant

cupid

The singleton instance of the Cupid type.

AstroBase.Bodies.cyllene — Constant

cyllene

The singleton instance of the Cyllene type.

AstroBase.Bodies.daphnis — Constant

daphnis

The singleton instance of the Daphnis type.

AstroBase.Bodies.davida — Constant

davida

The singleton instance of the Davida type.

AstroBase.Bodies.deimos — Constant

deimos

The singleton instance of the Deimos type.

AstroBase.Bodies.desdemona — Constant

desdemona

The singleton instance of the Desdemona type.

AstroBase.Bodies.despina — Constant

despina

The singleton instance of the Despina type.

AstroBase.Bodies.dia — Constant

dia

The singleton instance of the Dia type.

AstroBase.Bodies.dione — Constant

dione

The singleton instance of the Dione type.

AstroBase.Bodies.earth — Constant

earth

The singleton instance of the Earth type.

AstroBase.Bodies.earth_barycenter — Constant

earth_barycenter

The singleton instance of the EarthBarycenter type.

AstroBase.Bodies.elara — Constant

elara

The singleton instance of the Elara type.

AstroBase.Bodies.enceladus — Constant

enceladus

The singleton instance of the Enceladus type.

AstroBase.Bodies.epimetheus — Constant

epimetheus

The singleton instance of the Epimetheus type.

AstroBase.Bodies.erinome — Constant

erinome

The singleton instance of the Erinome type.

AstroBase.Bodies.eros — Constant

eros

The singleton instance of the Eros type.

AstroBase.Bodies.erriapus — Constant

erriapus

The singleton instance of the Erriapus type.

AstroBase.Bodies.euanthe — Constant

euanthe

The singleton instance of the Euanthe type.

AstroBase.Bodies.eukelade — Constant

eukelade

The singleton instance of the Eukelade type.

AstroBase.Bodies.euporie — Constant

euporie

The singleton instance of the Euporie type.

AstroBase.Bodies.europa — Constant

europa

The singleton instance of the Europa type.

AstroBase.Bodies.eurydome — Constant

eurydome

The singleton instance of the Eurydome type.

AstroBase.Bodies.farbauti — Constant

farbauti

The singleton instance of the Farbauti type.

AstroBase.Bodies.fenrir — Constant

fenrir

The singleton instance of the Fenrir type.

AstroBase.Bodies.ferdinand — Constant

ferdinand

The singleton instance of the Ferdinand type.

AstroBase.Bodies.fornjot — Constant

fornjot

The singleton instance of the Fornjot type.

AstroBase.Bodies.francisco — Constant

francisco

The singleton instance of the Francisco type.

AstroBase.Bodies.galatea — Constant

galatea

The singleton instance of the Galatea type.

AstroBase.Bodies.ganymede — Constant

ganymede

The singleton instance of the Ganymede type.

AstroBase.Bodies.gaspra — Constant

gaspra

The singleton instance of the Gaspra type.

AstroBase.Bodies.greip — Constant

greip

The singleton instance of the Greip type.

AstroBase.Bodies.halimede — Constant

halimede

The singleton instance of the Halimede type.

AstroBase.Bodies.harpalyke — Constant

harpalyke

The singleton instance of the Harpalyke type.

AstroBase.Bodies.hati — Constant

hati

The singleton instance of the Hati type.

AstroBase.Bodies.hegemone — Constant

hegemone

The singleton instance of the Hegemone type.

AstroBase.Bodies.helene — Constant

helene

The singleton instance of the Helene type.

AstroBase.Bodies.helike — Constant

helike

The singleton instance of the Helike type.

AstroBase.Bodies.hermippe — Constant

hermippe

The singleton instance of the Hermippe type.

AstroBase.Bodies.herse — Constant

herse

The singleton instance of the Herse type.

AstroBase.Bodies.himalia — Constant

himalia

The singleton instance of the Himalia type.

AstroBase.Bodies.hydra — Constant

hydra

The singleton instance of the Hydra type.

AstroBase.Bodies.hyperion — Constant

hyperion

The singleton instance of the Hyperion type.

AstroBase.Bodies.hyrrokkin — Constant

hyrrokkin

The singleton instance of the Hyrrokkin type.

AstroBase.Bodies.iapetus — Constant

iapetus

The singleton instance of the Iapetus type.

AstroBase.Bodies.ida — Constant

ida

The singleton instance of the Ida type.

AstroBase.Bodies.ijiraq — Constant

ijiraq

The singleton instance of the Ijiraq type.

AstroBase.Bodies.io — Constant

io

The singleton instance of the Io type.

AstroBase.Bodies.iocaste — Constant

iocaste

The singleton instance of the Iocaste type.

AstroBase.Bodies.isonoe — Constant

isonoe

The singleton instance of the Isonoe type.

AstroBase.Bodies.itokawa — Constant

itokawa

The singleton instance of the Itokawa type.

AstroBase.Bodies.janus — Constant

janus

The singleton instance of the Janus type.

AstroBase.Bodies.jarnsaxa — Constant

jarnsaxa

The singleton instance of the Jarnsaxa type.

AstroBase.Bodies.juliet — Constant

juliet

The singleton instance of the Juliet type.

AstroBase.Bodies.jupiter — Constant

jupiter

The singleton instance of the Jupiter type.

AstroBase.Bodies.jupiter_barycenter — Constant

jupiter_barycenter

The singleton instance of the JupiterBarycenter type.

AstroBase.Bodies.kale — Constant

kale

The singleton instance of the Kale type.

AstroBase.Bodies.kallichore — Constant

kallichore

The singleton instance of the Kallichore type.

AstroBase.Bodies.kalyke — Constant

kalyke

The singleton instance of the Kalyke type.

AstroBase.Bodies.kari — Constant

kari

The singleton instance of the Kari type.

AstroBase.Bodies.kerberos — Constant

kerberos

The singleton instance of the Kerberos type.

AstroBase.Bodies.kiviuq — Constant

kiviuq

The singleton instance of the Kiviuq type.

AstroBase.Bodies.kore — Constant

kore

The singleton instance of the Kore type.

AstroBase.Bodies.laomedeia — Constant

laomedeia

The singleton instance of the Laomedeia type.

AstroBase.Bodies.larissa — Constant

larissa

The singleton instance of the Larissa type.

AstroBase.Bodies.leda — Constant

leda

The singleton instance of the Leda type.

AstroBase.Bodies.loge — Constant

loge

The singleton instance of the Loge type.

AstroBase.Bodies.luna — Constant

luna

The singleton instance of the Luna type.

AstroBase.Bodies.lutetia — Constant

lutetia

The singleton instance of the Lutetia type.

AstroBase.Bodies.lysithea — Constant

lysithea

The singleton instance of the Lysithea type.

AstroBase.Bodies.mab — Constant

mab

The singleton instance of the Mab type.

AstroBase.Bodies.magaclite — Constant

magaclite

The singleton instance of the Magaclite type.

AstroBase.Bodies.margaret — Constant

margaret

The singleton instance of the Margaret type.

AstroBase.Bodies.mars — Constant

mars

The singleton instance of the Mars type.

AstroBase.Bodies.mars_barycenter — Constant

mars_barycenter

The singleton instance of the MarsBarycenter type.

AstroBase.Bodies.mercury — Constant

mercury

The singleton instance of the Mercury type.

AstroBase.Bodies.mercury_barycenter — Constant

mercury_barycenter

The singleton instance of the MercuryBarycenter type.

AstroBase.Bodies.methone — Constant

methone

The singleton instance of the Methone type.

AstroBase.Bodies.metis — Constant

metis

The singleton instance of the Metis type.

AstroBase.Bodies.mimas — Constant

mimas

The singleton instance of the Mimas type.

AstroBase.Bodies.miranda — Constant

miranda

The singleton instance of the Miranda type.

AstroBase.Bodies.mneme — Constant

mneme

The singleton instance of the Mneme type.

AstroBase.Bodies.moon — Constant

moon

moon is an alias for luna.

AstroBase.Bodies.mundilfari — Constant

mundilfari

The singleton instance of the Mundilfari type.

AstroBase.Bodies.naiad — Constant

naiad

The singleton instance of the Naiad type.

AstroBase.Bodies.narvi — Constant

narvi

The singleton instance of the Narvi type.

AstroBase.Bodies.neptune — Constant

neptune

The singleton instance of the Neptune type.

AstroBase.Bodies.neptune_barycenter — Constant

neptune_barycenter

The singleton instance of the NeptuneBarycenter type.

AstroBase.Bodies.nereid — Constant

nereid

The singleton instance of the Nereid type.

AstroBase.Bodies.neso — Constant

neso

The singleton instance of the Neso type.

AstroBase.Bodies.nix — Constant

nix

The singleton instance of the Nix type.

AstroBase.Bodies.oberon — Constant

oberon

The singleton instance of the Oberon type.

AstroBase.Bodies.ophelia — Constant

ophelia

The singleton instance of the Ophelia type.

AstroBase.Bodies.orthosie — Constant

orthosie

The singleton instance of the Orthosie type.

AstroBase.Bodies.paaliaq — Constant

paaliaq

The singleton instance of the Paaliaq type.

AstroBase.Bodies.pallas — Constant

pallas

The singleton instance of the Pallas type.

AstroBase.Bodies.pallene — Constant

pallene

The singleton instance of the Pallene type.

AstroBase.Bodies.pan — Constant

pan

The singleton instance of the Pan type.

AstroBase.Bodies.pandora — Constant

pandora

The singleton instance of the Pandora type.

AstroBase.Bodies.pasiphae — Constant

pasiphae

The singleton instance of the Pasiphae type.

AstroBase.Bodies.pasithee — Constant

pasithee

The singleton instance of the Pasithee type.

AstroBase.Bodies.perdita — Constant

perdita

The singleton instance of the Perdita type.

AstroBase.Bodies.phobos — Constant

phobos

The singleton instance of the Phobos type.

AstroBase.Bodies.phoebe — Constant

phoebe

The singleton instance of the Phoebe type.

AstroBase.Bodies.pluto — Constant

pluto

The singleton instance of the Pluto type.

AstroBase.Bodies.pluto_barycenter — Constant

pluto_barycenter

The singleton instance of the PlutoBarycenter type.

AstroBase.Bodies.polydeuces — Constant

polydeuces

The singleton instance of the Polydeuces type.

AstroBase.Bodies.portia — Constant

portia

The singleton instance of the Portia type.

AstroBase.Bodies.praxidike — Constant

praxidike

The singleton instance of the Praxidike type.

AstroBase.Bodies.prometheus — Constant

prometheus

The singleton instance of the Prometheus type.

AstroBase.Bodies.prospero — Constant

prospero

The singleton instance of the Prospero type.

AstroBase.Bodies.proteus — Constant

proteus

The singleton instance of the Proteus type.

AstroBase.Bodies.psamathe — Constant

psamathe

The singleton instance of the Psamathe type.

AstroBase.Bodies.psyche — Constant

psyche

The singleton instance of the Psyche type.

AstroBase.Bodies.puck — Constant

puck

The singleton instance of the Puck type.

AstroBase.Bodies.rhea — Constant

rhea

The singleton instance of the Rhea type.

AstroBase.Bodies.rosalind — Constant

rosalind

The singleton instance of the Rosalind type.

AstroBase.Bodies.sao — Constant

sao

The singleton instance of the Sao type.

AstroBase.Bodies.saturn — Constant

saturn

The singleton instance of the Saturn type.

AstroBase.Bodies.saturn_barycenter — Constant

saturn_barycenter

The singleton instance of the SaturnBarycenter type.

AstroBase.Bodies.setebos — Constant

setebos

The singleton instance of the Setebos type.

AstroBase.Bodies.siarnaq — Constant

siarnaq

The singleton instance of the Siarnaq type.

AstroBase.Bodies.sinope — Constant

sinope

The singleton instance of the Sinope type.

AstroBase.Bodies.skathi — Constant

skathi

The singleton instance of the Skathi type.

AstroBase.Bodies.skoll — Constant

skoll

The singleton instance of the Skoll type.

AstroBase.Bodies.sponde — Constant

sponde

The singleton instance of the Sponde type.

AstroBase.Bodies.ssb — Constant

ssb

The singleton instance of the Ssb type.

AstroBase.Bodies.steins — Constant

steins

The singleton instance of the Steins type.

AstroBase.Bodies.stephano — Constant

stephano

The singleton instance of the Stephano type.

AstroBase.Bodies.styx — Constant

styx

The singleton instance of the Styx type.

AstroBase.Bodies.sun — Constant

sun

The singleton instance of the Sun type.

AstroBase.Bodies.surtur — Constant

surtur

The singleton instance of the Surtur type.

AstroBase.Bodies.suttungr — Constant

suttungr

The singleton instance of the Suttungr type.

AstroBase.Bodies.sycorax — Constant

sycorax

The singleton instance of the Sycorax type.

AstroBase.Bodies.tarqeq — Constant

tarqeq

The singleton instance of the Tarqeq type.

AstroBase.Bodies.tarvos — Constant

tarvos

The singleton instance of the Tarvos type.

AstroBase.Bodies.taygete — Constant

taygete

The singleton instance of the Taygete type.

AstroBase.Bodies.telesto — Constant

telesto

The singleton instance of the Telesto type.

AstroBase.Bodies.tempel1 — Constant

tempel1

The singleton instance of the Tempel1 type.

AstroBase.Bodies.tethys — Constant

tethys

The singleton instance of the Tethys type.

AstroBase.Bodies.thalassa — Constant

thalassa

The singleton instance of the Thalassa type.

AstroBase.Bodies.thebe — Constant

thebe

The singleton instance of the Thebe type.

AstroBase.Bodies.thelxinoe — Constant

thelxinoe

The singleton instance of the Thelxinoe type.

AstroBase.Bodies.themisto — Constant

themisto

The singleton instance of the Themisto type.

AstroBase.Bodies.thrymr — Constant

thrymr

The singleton instance of the Thrymr type.

AstroBase.Bodies.thyone — Constant

thyone

The singleton instance of the Thyone type.

AstroBase.Bodies.titan — Constant

titan

The singleton instance of the Titan type.

AstroBase.Bodies.titania — Constant

titania

The singleton instance of the Titania type.

AstroBase.Bodies.trinculo — Constant

trinculo

The singleton instance of the Trinculo type.

AstroBase.Bodies.triton — Constant

triton

The singleton instance of the Triton type.

AstroBase.Bodies.umbriel — Constant

umbriel

The singleton instance of the Umbriel type.

AstroBase.Bodies.uranus — Constant

uranus

The singleton instance of the Uranus type.

AstroBase.Bodies.uranus_barycenter — Constant

uranus_barycenter

The singleton instance of the UranusBarycenter type.

AstroBase.Bodies.venus — Constant

venus

The singleton instance of the Venus type.

AstroBase.Bodies.venus_barycenter — Constant

venus_barycenter

The singleton instance of the VenusBarycenter type.

AstroBase.Bodies.vesta — Constant

vesta

The singleton instance of the Vesta type.

AstroBase.Bodies.ymir — Constant

ymir

The singleton instance of the Ymir type.

AstroBase.Bodies.Adrastea — Type

Adrastea <: NaturalSatellite

A type representing Adrastea.

AstroBase.Bodies.Aegaeon — Type

Aegaeon <: NaturalSatellite

A type representing Aegaeon.

AstroBase.Bodies.Aegir — Type

Aegir <: NaturalSatellite

A type representing Aegir.

AstroBase.Bodies.Aitne — Type

Aitne <: NaturalSatellite

A type representing Aitne.

AstroBase.Bodies.Albiorix — Type

Albiorix <: NaturalSatellite

A type representing Albiorix.

AstroBase.Bodies.Amalthea — Type

Amalthea <: NaturalSatellite

A type representing Amalthea.

AstroBase.Bodies.Ananke — Type

Ananke <: NaturalSatellite

A type representing Ananke.

AstroBase.Bodies.Anthe — Type

Anthe <: NaturalSatellite

A type representing Anthe.

AstroBase.Bodies.Aoede — Type

Aoede <: NaturalSatellite

A type representing Aoede.

AstroBase.Bodies.Arche — Type

Arche <: NaturalSatellite

A type representing Arche.

AstroBase.Bodies.Ariel — Type

Ariel <: NaturalSatellite

A type representing Ariel.

AstroBase.Bodies.Atlas — Type

Atlas <: NaturalSatellite

A type representing Atlas.

AstroBase.Bodies.Autonoe — Type

Autonoe <: NaturalSatellite

A type representing Autonoe.

AstroBase.Bodies.Barycenter — Type

Barycenter <: CelestialBody

Abstract supertype for representing the barycenters of the solar system and planetary systems as pseudo-bodies.

AstroBase.Bodies.Bebhionn — Type

Bebhionn <: NaturalSatellite

A type representing Bebhionn.

AstroBase.Bodies.Belinda — Type

Belinda <: NaturalSatellite

A type representing Belinda.

AstroBase.Bodies.Bergelmir — Type

Bergelmir <: NaturalSatellite

A type representing Bergelmir.

AstroBase.Bodies.Bestla — Type

Bestla <: NaturalSatellite

A type representing Bestla.

AstroBase.Bodies.Bianca — Type

Bianca <: NaturalSatellite

A type representing Bianca.

AstroBase.Bodies.Borrelly — Type

Borrelly <: MinorBody

A type representing Borrelly.

AstroBase.Bodies.Caliban — Type

Caliban <: NaturalSatellite

A type representing Caliban.

AstroBase.Bodies.Callirrhoe — Type

Callirrhoe <: NaturalSatellite

A type representing Callirrhoe.

AstroBase.Bodies.Callisto — Type

Callisto <: NaturalSatellite

A type representing Callisto.

AstroBase.Bodies.Calypso — Type

Calypso <: NaturalSatellite

A type representing Calypso.

AstroBase.Bodies.Carme — Type

Carme <: NaturalSatellite

A type representing Carme.

AstroBase.Bodies.Carpo — Type

Carpo <: NaturalSatellite

A type representing Carpo.

AstroBase.Bodies.CelestialBody — Type

CelestialBody

Abstract supertype for all celestial bodies and pseudo-bodies.

AstroBase.Bodies.Ceres — Type

Ceres <: MinorBody

A type representing Ceres.

AstroBase.Bodies.Chaldene — Type

Chaldene <: NaturalSatellite

A type representing Chaldene.

AstroBase.Bodies.Charon — Type

Charon <: NaturalSatellite

A type representing Charon.

AstroBase.Bodies.Cordelia — Type

Cordelia <: NaturalSatellite

A type representing Cordelia.

AstroBase.Bodies.Cressida — Type

Cressida <: NaturalSatellite

A type representing Cressida.

AstroBase.Bodies.Cupid — Type

Cupid <: NaturalSatellite

A type representing Cupid.

AstroBase.Bodies.Cyllene — Type

Cyllene <: NaturalSatellite

A type representing Cyllene.

AstroBase.Bodies.Daphnis — Type

Daphnis <: NaturalSatellite

A type representing Daphnis.

AstroBase.Bodies.Davida — Type

Davida <: MinorBody

A type representing Davida.

AstroBase.Bodies.Deimos — Type

Deimos <: NaturalSatellite

A type representing Deimos.

AstroBase.Bodies.Desdemona — Type

Desdemona <: NaturalSatellite

A type representing Desdemona.

AstroBase.Bodies.Despina — Type

Despina <: NaturalSatellite

A type representing Despina.

AstroBase.Bodies.Dia — Type

Dia <: NaturalSatellite

A type representing Dia.

AstroBase.Bodies.Dione — Type

Dione <: NaturalSatellite

A type representing Dione.

AstroBase.Bodies.Earth — Type

Earth <: Planet

A type representing Earth.

AstroBase.Bodies.EarthBarycenter — Type

EarthBarycenter <: Barycenter

A type representing the Earth Barycenter.

AstroBase.Bodies.Elara — Type

Elara <: NaturalSatellite

A type representing Elara.

AstroBase.Bodies.Enceladus — Type

Enceladus <: NaturalSatellite

A type representing Enceladus.

AstroBase.Bodies.Epimetheus — Type

Epimetheus <: NaturalSatellite

A type representing Epimetheus.

AstroBase.Bodies.Erinome — Type

Erinome <: NaturalSatellite

A type representing Erinome.

AstroBase.Bodies.Eros — Type

Eros <: MinorBody

A type representing Eros.

AstroBase.Bodies.Erriapus — Type

Erriapus <: NaturalSatellite

A type representing Erriapus.

AstroBase.Bodies.Euanthe — Type

Euanthe <: NaturalSatellite

A type representing Euanthe.

AstroBase.Bodies.Eukelade — Type

Eukelade <: NaturalSatellite

A type representing Eukelade.

AstroBase.Bodies.Euporie — Type

Euporie <: NaturalSatellite

A type representing Euporie.

AstroBase.Bodies.Europa — Type

Europa <: NaturalSatellite

A type representing Europa.

AstroBase.Bodies.Eurydome — Type

Eurydome <: NaturalSatellite

A type representing Eurydome.

AstroBase.Bodies.Farbauti — Type

Farbauti <: NaturalSatellite

A type representing Farbauti.

AstroBase.Bodies.Fenrir — Type

Fenrir <: NaturalSatellite

A type representing Fenrir.

AstroBase.Bodies.Ferdinand — Type

Ferdinand <: NaturalSatellite

A type representing Ferdinand.

AstroBase.Bodies.Fornjot — Type

Fornjot <: NaturalSatellite

A type representing Fornjot.

AstroBase.Bodies.Francisco — Type

Francisco <: NaturalSatellite

A type representing Francisco.

AstroBase.Bodies.Galatea — Type

Galatea <: NaturalSatellite

A type representing Galatea.

AstroBase.Bodies.Ganymede — Type

Ganymede <: NaturalSatellite

A type representing Ganymede.

AstroBase.Bodies.Gaspra — Type

Gaspra <: MinorBody

A type representing Gaspra.

AstroBase.Bodies.Greip — Type

Greip <: NaturalSatellite

A type representing Greip.

AstroBase.Bodies.Halimede — Type

Halimede <: NaturalSatellite

A type representing Halimede.

AstroBase.Bodies.Harpalyke — Type

Harpalyke <: NaturalSatellite

A type representing Harpalyke.

AstroBase.Bodies.Hati — Type

Hati <: NaturalSatellite

A type representing Hati.

AstroBase.Bodies.Hegemone — Type

Hegemone <: NaturalSatellite

A type representing Hegemone.

AstroBase.Bodies.Helene — Type

Helene <: NaturalSatellite

A type representing Helene.

AstroBase.Bodies.Helike — Type

Helike <: NaturalSatellite

A type representing Helike.

AstroBase.Bodies.Hermippe — Type

Hermippe <: NaturalSatellite

A type representing Hermippe.

AstroBase.Bodies.Herse — Type

Herse <: NaturalSatellite

A type representing Herse.

AstroBase.Bodies.Himalia — Type

Himalia <: NaturalSatellite

A type representing Himalia.

AstroBase.Bodies.Hydra — Type

Hydra <: NaturalSatellite

A type representing Hydra.

AstroBase.Bodies.Hyperion — Type

Hyperion <: NaturalSatellite

A type representing Hyperion.

AstroBase.Bodies.Hyrrokkin — Type

Hyrrokkin <: NaturalSatellite

A type representing Hyrrokkin.

AstroBase.Bodies.Iapetus — Type

Iapetus <: NaturalSatellite

A type representing Iapetus.

AstroBase.Bodies.Ida — Type

Ida <: MinorBody

A type representing Ida.

AstroBase.Bodies.Ijiraq — Type

Ijiraq <: NaturalSatellite

A type representing Ijiraq.

AstroBase.Bodies.Io — Type

Io <: NaturalSatellite

A type representing Io.

AstroBase.Bodies.Iocaste — Type

Iocaste <: NaturalSatellite

A type representing Iocaste.

AstroBase.Bodies.Isonoe — Type

Isonoe <: NaturalSatellite

A type representing Isonoe.

AstroBase.Bodies.Itokawa — Type

Itokawa <: MinorBody

A type representing Itokawa.

AstroBase.Bodies.Janus — Type

Janus <: NaturalSatellite

A type representing Janus.

AstroBase.Bodies.Jarnsaxa — Type

Jarnsaxa <: NaturalSatellite

A type representing Jarnsaxa.

AstroBase.Bodies.Juliet — Type

Juliet <: NaturalSatellite

A type representing Juliet.

AstroBase.Bodies.Jupiter — Type

Jupiter <: Planet

A type representing Jupiter.

AstroBase.Bodies.JupiterBarycenter — Type

JupiterBarycenter <: Barycenter

A type representing the Jupiter Barycenter.

AstroBase.Bodies.Kale — Type

Kale <: NaturalSatellite

A type representing Kale.

AstroBase.Bodies.Kallichore — Type

Kallichore <: NaturalSatellite

A type representing Kallichore.

AstroBase.Bodies.Kalyke — Type

Kalyke <: NaturalSatellite

A type representing Kalyke.

AstroBase.Bodies.Kari — Type

Kari <: NaturalSatellite

A type representing Kari.

AstroBase.Bodies.Kerberos — Type

Kerberos <: NaturalSatellite

A type representing Kerberos.

AstroBase.Bodies.Kiviuq — Type

Kiviuq <: NaturalSatellite

A type representing Kiviuq.

AstroBase.Bodies.Kore — Type

Kore <: NaturalSatellite

A type representing Kore.

AstroBase.Bodies.Laomedeia — Type

Laomedeia <: NaturalSatellite

A type representing Laomedeia.

AstroBase.Bodies.Larissa — Type

Larissa <: NaturalSatellite

A type representing Larissa.

AstroBase.Bodies.Leda — Type

Leda <: NaturalSatellite

A type representing Leda.

AstroBase.Bodies.Loge — Type

Loge <: NaturalSatellite

A type representing Loge.

AstroBase.Bodies.Luna — Type

Luna <: NaturalSatellite

A type representing Luna.

AstroBase.Bodies.Lutetia — Type

Lutetia <: MinorBody

A type representing Lutetia.

AstroBase.Bodies.Lysithea — Type

Lysithea <: NaturalSatellite

A type representing Lysithea.

AstroBase.Bodies.Mab — Type

Mab <: NaturalSatellite

A type representing Mab.

AstroBase.Bodies.Magaclite — Type

Magaclite <: NaturalSatellite

A type representing Magaclite.

AstroBase.Bodies.Margaret — Type

Margaret <: NaturalSatellite

A type representing Margaret.

AstroBase.Bodies.Mars — Type

Mars <: Planet

A type representing Mars.

AstroBase.Bodies.MarsBarycenter — Type

MarsBarycenter <: Barycenter

A type representing the Mars Barycenter.

AstroBase.Bodies.Mercury — Type

Mercury <: Planet

A type representing Mercury.

AstroBase.Bodies.MercuryBarycenter — Type

MercuryBarycenter <: Barycenter

A type representing the Mercury Barycenter.

AstroBase.Bodies.Methone — Type

Methone <: NaturalSatellite

A type representing Methone.

AstroBase.Bodies.Metis — Type

Metis <: NaturalSatellite

A type representing Metis.

AstroBase.Bodies.Mimas — Type

Mimas <: NaturalSatellite

A type representing Mimas.

AstroBase.Bodies.MinorBody — Type

MinorBody <: CelestialBody

Abstract supertype for minor solar system bodies.

AstroBase.Bodies.Miranda — Type

Miranda <: NaturalSatellite

A type representing Miranda.

AstroBase.Bodies.Mneme — Type

Mneme <: NaturalSatellite

A type representing Mneme.

AstroBase.Bodies.Mundilfari — Type

Mundilfari <: NaturalSatellite

A type representing Mundilfari.

AstroBase.Bodies.NAIFId — Type

NAIFId

An integer code for identifying celestial bodies and other objects in space.

References

NASA NAIF

AstroBase.Bodies.Naiad — Type

Naiad <: NaturalSatellite

A type representing Naiad.

AstroBase.Bodies.Narvi — Type

Narvi <: NaturalSatellite

A type representing Narvi.

AstroBase.Bodies.NaturalSatellite — Type

NaturalSatellite <: CelestialBody

Abstract supertype for natural satellites (moons).

AstroBase.Bodies.Neptune — Type

Neptune <: Planet

A type representing Neptune.

AstroBase.Bodies.NeptuneBarycenter — Type

NeptuneBarycenter <: Barycenter

A type representing the Neptune Barycenter.

AstroBase.Bodies.Nereid — Type

Nereid <: NaturalSatellite

A type representing Nereid.

AstroBase.Bodies.Neso — Type

Neso <: NaturalSatellite

A type representing Neso.

AstroBase.Bodies.Nix — Type

Nix <: NaturalSatellite

A type representing Nix.

AstroBase.Bodies.Oberon — Type

Oberon <: NaturalSatellite

A type representing Oberon.

AstroBase.Bodies.Ophelia — Type

Ophelia <: NaturalSatellite

A type representing Ophelia.

AstroBase.Bodies.Orthosie — Type

Orthosie <: NaturalSatellite

A type representing Orthosie.

AstroBase.Bodies.Paaliaq — Type

Paaliaq <: NaturalSatellite

A type representing Paaliaq.

AstroBase.Bodies.Pallas — Type

Pallas <: MinorBody

A type representing Pallas.

AstroBase.Bodies.Pallene — Type

Pallene <: NaturalSatellite

A type representing Pallene.

AstroBase.Bodies.Pan — Type

Pan <: NaturalSatellite

A type representing Pan.

AstroBase.Bodies.Pandora — Type

Pandora <: NaturalSatellite

A type representing Pandora.

AstroBase.Bodies.Pasiphae — Type

Pasiphae <: NaturalSatellite

A type representing Pasiphae.

AstroBase.Bodies.Pasithee — Type

Pasithee <: NaturalSatellite

A type representing Pasithee.

AstroBase.Bodies.Perdita — Type

Perdita <: NaturalSatellite

A type representing Perdita.

AstroBase.Bodies.Phobos — Type

Phobos <: NaturalSatellite

A type representing Phobos.

AstroBase.Bodies.Phoebe — Type

Phoebe <: NaturalSatellite

A type representing Phoebe.

AstroBase.Bodies.Planet — Type

Planet <: CelestialBody

Abstract supertype for planets.

AstroBase.Bodies.Pluto — Type

Pluto <: MinorBody

A type representing Pluto.

AstroBase.Bodies.PlutoBarycenter — Type

PlutoBarycenter <: Barycenter

A type representing the Pluto Barycenter.

AstroBase.Bodies.Polydeuces — Type

Polydeuces <: NaturalSatellite

A type representing Polydeuces.

AstroBase.Bodies.Portia — Type

Portia <: NaturalSatellite

A type representing Portia.

AstroBase.Bodies.Praxidike — Type

Praxidike <: NaturalSatellite

A type representing Praxidike.

AstroBase.Bodies.Prometheus — Type

Prometheus <: NaturalSatellite

A type representing Prometheus.

AstroBase.Bodies.Prospero — Type

Prospero <: NaturalSatellite

A type representing Prospero.

AstroBase.Bodies.Proteus — Type

Proteus <: NaturalSatellite

A type representing Proteus.

AstroBase.Bodies.Psamathe — Type

Psamathe <: NaturalSatellite

A type representing Psamathe.

AstroBase.Bodies.Psyche — Type

Psyche <: MinorBody

A type representing Psyche.

AstroBase.Bodies.Puck — Type

Puck <: NaturalSatellite

A type representing Puck.

AstroBase.Bodies.Rhea — Type

Rhea <: NaturalSatellite

A type representing Rhea.

AstroBase.Bodies.Rosalind — Type

Rosalind <: NaturalSatellite

A type representing Rosalind.

AstroBase.Bodies.Sao — Type

Sao <: NaturalSatellite

A type representing Sao.

AstroBase.Bodies.Saturn — Type

Saturn <: Planet

A type representing Saturn.

AstroBase.Bodies.SaturnBarycenter — Type

SaturnBarycenter <: Barycenter

A type representing the Saturn Barycenter.

AstroBase.Bodies.Setebos — Type

Setebos <: NaturalSatellite

A type representing Setebos.

AstroBase.Bodies.Siarnaq — Type

Siarnaq <: NaturalSatellite

A type representing Siarnaq.

AstroBase.Bodies.Sinope — Type

Sinope <: NaturalSatellite

A type representing Sinope.

AstroBase.Bodies.Skathi — Type

Skathi <: NaturalSatellite

A type representing Skathi.

AstroBase.Bodies.Skoll — Type

Skoll <: NaturalSatellite

A type representing Skoll.

AstroBase.Bodies.SolarSystemBarycenter — Type

Ssb <: Barycenter

A type representing the Solar System Barycenter.

AstroBase.Bodies.Sponde — Type

Sponde <: NaturalSatellite

A type representing Sponde.

AstroBase.Bodies.Steins — Type

Steins <: MinorBody

A type representing Steins.

AstroBase.Bodies.Stephano — Type

Stephano <: NaturalSatellite

A type representing Stephano.

AstroBase.Bodies.Styx — Type

Styx <: NaturalSatellite

A type representing Styx.

AstroBase.Bodies.Sun — Type

Sun <: CelestialBody

A type representing Sun.

AstroBase.Bodies.Surtur — Type

Surtur <: NaturalSatellite

A type representing Surtur.

AstroBase.Bodies.Suttungr — Type

Suttungr <: NaturalSatellite

A type representing Suttungr.

AstroBase.Bodies.Sycorax — Type

Sycorax <: NaturalSatellite

A type representing Sycorax.

AstroBase.Bodies.Tarqeq — Type

Tarqeq <: NaturalSatellite

A type representing Tarqeq.

AstroBase.Bodies.Tarvos — Type

Tarvos <: NaturalSatellite

A type representing Tarvos.

AstroBase.Bodies.Taygete — Type

Taygete <: NaturalSatellite

A type representing Taygete.

AstroBase.Bodies.Telesto — Type

Telesto <: NaturalSatellite

A type representing Telesto.

AstroBase.Bodies.Tempel1 — Type

Tempel1 <: MinorBody

A type representing Tempel 1.

AstroBase.Bodies.Tethys — Type

Tethys <: NaturalSatellite

A type representing Tethys.

AstroBase.Bodies.Thalassa — Type

Thalassa <: NaturalSatellite

A type representing Thalassa.

AstroBase.Bodies.Thebe — Type

Thebe <: NaturalSatellite

A type representing Thebe.

AstroBase.Bodies.Thelxinoe — Type

Thelxinoe <: NaturalSatellite

A type representing Thelxinoe.

AstroBase.Bodies.Themisto — Type

Themisto <: NaturalSatellite

A type representing Themisto.

AstroBase.Bodies.Thrymr — Type

Thrymr <: NaturalSatellite

A type representing Thrymr.

AstroBase.Bodies.Thyone — Type

Thyone <: NaturalSatellite

A type representing Thyone.

AstroBase.Bodies.Titan — Type

Titan <: NaturalSatellite

A type representing Titan.

AstroBase.Bodies.Titania — Type

Titania <: NaturalSatellite

A type representing Titania.

AstroBase.Bodies.Trinculo — Type

Trinculo <: NaturalSatellite

A type representing Trinculo.

AstroBase.Bodies.Triton — Type

Triton <: NaturalSatellite

A type representing Triton.

AstroBase.Bodies.Umbriel — Type

Umbriel <: NaturalSatellite

A type representing Umbriel.

AstroBase.Bodies.Uranus — Type

Uranus <: Planet

A type representing Uranus.

AstroBase.Bodies.UranusBarycenter — Type

UranusBarycenter <: Barycenter

A type representing the Uranus Barycenter.

AstroBase.Bodies.Venus — Type

Venus <: Planet

A type representing Venus.

AstroBase.Bodies.VenusBarycenter — Type

VenusBarycenter <: Barycenter

A type representing the Venus Barycenter.

AstroBase.Bodies.Vesta — Type

Vesta <: MinorBody

A type representing Vesta.

AstroBase.Bodies.Ymir — Type

Ymir <: NaturalSatellite

A type representing Ymir.

AstroBase.Bodies.along_orbit_radius — Function

along_orbit_radius([T=Float64,] body::CelestialBody)

Return the along-orbit radius of body in km.

Example

julia> along_orbit_radius(earth)
6378.1366

References

Archinal, Brent Allen, et al. "Report of the IAU working group on cartographic coordinates and rotational elements: 2009." Celestial Mechanics and Dynamical Astronomy 109.2 (2011): 101-135.

AstroBase.Bodies.declination — Function

declination([[T=Float64,] PT=Float64,], body, ep)

Return the declination of the body-fixed frame of body w.r.t. the ICRF at epoch ep in rad.

Example

julia> declination(earth, TTEpoch(2000, 1, 1))
1.5707964598747588

References

Archinal, Brent Allen, et al. "Report of the IAU working group on cartographic coordinates and rotational elements: 2009." Celestial Mechanics and Dynamical Astronomy 109.2 (2011): 101-135.

AstroBase.Bodies.declination_rate — Function

declination_rate([[T=Float64,] PT=Float64,], body, ep)

Return the declination rate of change of the body-fixed frame of body w.r.t. the ICRF at epoch ep in rad/s.

Example

julia> declination_rate(earth, TTEpoch(2000, 1, 1))
-3.0805523657085508e-12

References

Archinal, Brent Allen, et al. "Report of the IAU working group on cartographic coordinates and rotational elements: 2009." Celestial Mechanics and Dynamical Astronomy 109.2 (2011): 101-135.

AstroBase.Bodies.ellipsoid — Method

ellipsoid([T=Float64,] body::CelestialBody)

Return the subplanetary, along-orbit, and polar radii of body in km which constitute its tri-axial ellipsoid.

Example

julia> ellipsoid(earth)
(6378.1366, 6378.1366, 6356.7519)

References

Archinal, Brent Allen, et al. "Report of the IAU working group on cartographic coordinates and rotational elements: 2009." Celestial Mechanics and Dynamical Astronomy 109.2 (2011): 101-135.

AstroBase.Bodies.equatorial_radius — Function

equatorial_radius([T=Float64,] body::CelestialBody)

Return the equatorial radius of body in km.

Note

The equatorial radius is only available for bodies which have identical subplanetary and along-orbit radii.

Example

julia> equatorial_radius(earth)
6378.1366

References

Archinal, Brent Allen, et al. "Report of the IAU working group on cartographic coordinates and rotational elements: 2009." Celestial Mechanics and Dynamical Astronomy 109.2 (2011): 101-135.

AstroBase.Bodies.euler_angles — Method

euler_angles([[T=Float64,] PT=Float64,], body, ep)

Return the orientation of the body-fixed frame of body w.r.t. the ICRF at epoch ep as a set of Euler angles in rad with rotation order ZXZ.

Example

julia> euler_angles(earth, TTEpoch(2000, 1, 1))
(1.5707964799442533, -1.3307986224120327e-7, 0.16849737156984945)

References

Archinal, Brent Allen, et al. "Report of the IAU working group on cartographic coordinates and rotational elements: 2009." Celestial Mechanics and Dynamical Astronomy 109.2 (2011): 101-135.

AstroBase.Bodies.euler_rates — Method

euler_rates([[T=Float64,] PT=Float64,], body, ep)

Return the rate of change of orientation of the body-fixed frame of body w.r.t. the ICRF at epoch ep as a set of Euler angle rates in rad/s with rotation order ZXZ.

Example

julia> euler_rates(earth, TTEpoch(2000, 1, 1))
(-3.545123997161905e-12, 3.0805523657085508e-12, 7.292115373194001e-5)

References

Archinal, Brent Allen, et al. "Report of the IAU working group on cartographic coordinates and rotational elements: 2009." Celestial Mechanics and Dynamical Astronomy 109.2 (2011): 101-135.

AstroBase.Bodies.from_naifid — Method

from_naifid(id)

Return a celestial body instance based on its NAIF ID code.

References

NASA NAIF

AstroBase.Bodies.grav_param — Function

grav_param([T=Float64,] body::CelestialBody)

Return the gravitational parameter $\mu = GM$ of body in km^3/s^2.

Example

julia> grav_param(earth)
398600.435436096

References

NASA NAIF

AstroBase.Bodies.mean_radius — Function

mean_radius([T=Float64,] body::CelestialBody)

Return the mean radius of body in km.

Example

julia> mean_radius(earth)
6371.008366666666

References

Archinal, Brent Allen, et al. "Report of the IAU working group on cartographic coordinates and rotational elements: 2009." Celestial Mechanics and Dynamical Astronomy 109.2 (2011): 101-135.

AstroBase.Bodies.naifid — Function

naifid(body::CelestialBody) -> NAIFId

Return the NAIF ID code for body.

References

NASA NAIF

AstroBase.Bodies.polar_radius — Function

polar_radius([T=Float64,] body::CelestialBody)

Return the polar radius of body in km.

Example

julia> polar_radius(earth)
6356.7519

References

Archinal, Brent Allen, et al. "Report of the IAU working group on cartographic coordinates and rotational elements: 2009." Celestial Mechanics and Dynamical Astronomy 109.2 (2011): 101-135.

AstroBase.Bodies.right_ascension — Function

right_ascension([[T=Float64,] PT=Float64,], body, ep)

Return the right ascension of the body-fixed frame of body w.r.t. the ICRF at epoch ep in rad.

Example

julia> right_ascension(earth, TTEpoch(2000, 1, 1))
1.5314935667739428e-7

References

Archinal, Brent Allen, et al. "Report of the IAU working group on cartographic coordinates and rotational elements: 2009." Celestial Mechanics and Dynamical Astronomy 109.2 (2011): 101-135.

AstroBase.Bodies.right_ascension_rate — Function

right_ascension_rate([[T=Float64,] PT=Float64,], body, ep)

Return the right ascension rate of change of the body-fixed frame of body w.r.t. the ICRF at epoch ep in rad/s.

Example

julia> right_ascension_rate(earth, TTEpoch(2000, 1, 1))
-3.545123997161905e-12

References

Archinal, Brent Allen, et al. "Report of the IAU working group on cartographic coordinates and rotational elements: 2009." Celestial Mechanics and Dynamical Astronomy 109.2 (2011): 101-135.

AstroBase.Bodies.rotation_angle — Function

rotation_angle([[T=Float64,] PT=Float64,], body, ep)

Return the rotation angle of the body-fixed frame of body w.r.t. the ICRF at epoch ep in rad.

Example

julia> rotation_angle(earth, TTEpoch(2000, 1, 1))
0.16849737156984945

References

Archinal, Brent Allen, et al. "Report of the IAU working group on cartographic coordinates and rotational elements: 2009." Celestial Mechanics and Dynamical Astronomy 109.2 (2011): 101-135.

AstroBase.Bodies.rotation_rate — Function

rotation_rate([[T=Float64,] PT=Float64,], body, ep)

Return the rotation rate of change of the body-fixed frame of body w.r.t. the ICRF at epoch ep in rad/s.

Example

julia> rotation_rate(earth, TTEpoch(2000, 1, 1))
7.292115373194001e-5

References

Archinal, Brent Allen, et al. "Report of the IAU working group on cartographic coordinates and rotational elements: 2009." Celestial Mechanics and Dynamical Astronomy 109.2 (2011): 101-135.

AstroBase.Bodies.subplanetary_radius — Function

subplanetary_radius([T=Float64,] body::CelestialBody)

Return the subplanetary radius of body in km.

Example

julia> subplanetary_radius(earth)
6378.1366

References

Archinal, Brent Allen, et al. "Report of the IAU working group on cartographic coordinates and rotational elements: 2009." Celestial Mechanics and Dynamical Astronomy 109.2 (2011): 101-135.

AstroBase.Bodies.@body — Macro

@body(name, id, supertype, type=Name, parent=nothing)

Define a new celestial body with NAIF id and supertype which is the singleton instance of type. Optionally, provide a parent (pseudo-)body.

Example

julia> @body rupert 42 MinorBody type=Persephone parent=ssb

julia> rupert isa Persephone
true

julia> typeof(rupert) <: MinorBody
true

julia> parent(rupert)
Solar System Barycenter

julia> naifid(rupert)
42