Earth Attitude

AstroBase.EarthAttitude.iau1980 — Constant

`iau1980`

The singleton instance of type IAU1980, representing the IAU 1980 family of models.

AstroBase.EarthAttitude.iau2000a — Constant

`iau2000a`

The singleton instance of type IAU2000A, representing the IAU 2000A family of models.

AstroBase.EarthAttitude.iau2000b — Constant

`iau2000b`

The singleton instance of type IAU2000B, representing the IAU 2000B family of models.

AstroBase.EarthAttitude.iau2006 — Constant

`iau2006`

The singleton instance of type IAU2006, representing the IAU 2006 family of models.

AstroBase.EarthAttitude.bias_precession_matrix_00 — Method

bias_precession_matrix_00(jd1, jd2)

Returns a tuple of array, rb (frame bias), rp (precession matrix), rbp (bias precession matrix) given 2 part Julian date (TT).

Example

julia> biasprecessionmatrix_00(2.4578265e6, 0.30434616919175345) (RotZYX(-7.07828e-8, -8.05622e-8, 3.30604e-8), [0.999991 0.00384587 0.00167106; -0.00384587 0.999993 -3.2343e-6; -0.00167106 -3.1924e-6 0.999999], [0.999991 0.00384594 0.00167098; -0.00384594 0.999993 -3.26705e-6; -0.00167098 -3.15946e-6 0.999999]) ```

AstroBase.EarthAttitude.c2tpe — Method

c2tpe(tta, ttb, uta, utb, dpsi, deps, xp, yp)

Returns the celestial to terrestrial matrix given two, 2 part Julian date (UT1, TT), the nutation and coordinates of the pole(radians).

Example

julia> c2tpe(2.4579405e6, 0.0, 2.4579405e6, -0.0007966009351851851, 0.1, 0.3, 0.5, 0.7)
3×3 Rotations.RotMatrix{3,Float64,9}:
  0.253143   0.777796  0.575284
 -0.676266  -0.28296   0.680146
  0.691797  -0.561219  0.454367

AstroBase.EarthAttitude.celestial_to_intermediate — Method

celestial_to_intermediate(x, y, s)

Returns celestial to intermediate-frame-of-date transformation matrix given the Celestial Intermediate Pole location (x, y and the CIO locator s).

Example

julia> celestial_to_intermediate(0.2, 0.2, 0.1)
3×3 RotZYZ{Float64}(0.785398, 0.286757, -0.885398):
  0.976728   0.0774803  0.2
 -0.11811    0.972651   0.2
 -0.179034  -0.218968   0.959166

AstroBase.EarthAttitude.celestial_to_intermediate_frame_of_date — Method

celestial_to_intermediate_frame_of_date(jd1, jd2, x, y)

Returns celestial to intermediate-frame-of-date matrix for a given 2 part Julian Date.

Example

julia> celestial_to_intermediate_frame_of_date(2.4578265e6, 0.30434616919175345, 0.2, 0.3)
3×3 Rotations.RotZYZ{Float64}(0.982794, 0.368863, -0.952794):
  0.977932    -0.0604048  0.2
 -0.00243134   0.953936   0.3
 -0.208909    -0.293866   0.932738

AstroBase.EarthAttitude.celestial_to_intermediate_matrix — Method

celestial_to_intermediate_matrix(jd1, jd2, rbpn)

Returns celestial-to-intermediate matrix for given 2 part Julian date and celestial-to-true matrix.

Example

julia> celestial_to_intermediate_matrix(2.4578265e6, 0.30434616919175345, rand(3,3))
3×3 Rotations.RotZYZ{Float64}(0.921674, 1.10597, -0.729284):
  0.732866  -0.413437  0.54035
 -0.13652    0.688686  0.712092
 -0.666536  -0.595636  0.448272

AstroBase.EarthAttitude.celestial_to_intermediate_matrix_a00 — Method

celestial_to_intermediate_matrix_a00(jd1, jd2)

Returns celestial-to-intermediate matrix for a given 2 part Julian date (TT).

Example

julia> celestial_to_intermediate_matrix_a00(2.4578265e6, 0.30434616919175345)
3×3 Rotations.RotZYZ{Float64}(-0.0240766, 0.00165637, 0.0240765):
  0.999999    6.21979e-8   0.00165589
  3.83181e-9  1.0         -3.98758e-5
 -0.00165589  3.98757e-5   0.999999

AstroBase.EarthAttitude.celestial_to_intermediate_matrix_b00 — Method

celestial_to_intermediate_matrix_b00(jd1, jd2)

Returns celestial-to-intermediate matrix for a given 2 part Julian date (TT).

Example

julia> celestial_to_intermediate_matrix_b00(2.4578265e6, 0.30434616919175345)
3×3 Rotations.RotZYZ{Float64}(-0.0240766, 0.00165637, 0.0240765):
0.999999    6.21979e-8   0.00165589
3.83181e-9  1.0         -3.98758e-5
-0.00165589  3.98757e-5   0.999999

AstroBase.EarthAttitude.celestial_to_terrestrial_a00 — Method

celestial_to_terrestrial_a00(tta, ttb, uta, utb, xp, yp)

Returns celestial to terrestrial matrix given two, 2 part Julian dates (UT1, TT) and coordinates of the pole.

Example

julia> celestial_to_terrestrial_a00(2.4579405e6, 0.0, 2.4579405e6, -0.0007966009351851851, 0.1, 0.3)
3×3 Rotations.RotMatrix{3,Float64,9}:
  0.235088    0.934797  0.266247
 -0.966879    0.196899  0.162409
  0.0993959  -0.295609  0.950124

AstroBase.EarthAttitude.celestial_to_terrestrial_b00 — Method

celestial_to_terrestrial_b00(tta, ttb, uta, utb, xp, yp)

Returns celestial to terrestrial matrix given two, 2 part Julian dates (UT1, TT) and coordinates of the pole.

Example

julia> celestial_to_terrestrial_b00(2.4579405e6, 0.0, 2.4579405e6, -0.0007966009351851851, 0.1, 0.3)
3×3 Rotations.RotMatrix{3,Float64,9}:
  0.235088    0.934797  0.266247
 -0.966879    0.196899  0.162409
  0.0993959  -0.295609  0.950124

AstroBase.EarthAttitude.celestial_to_terrestrial_matrix — Method

celestial_to_terrestrial_matrix(tta, ttb, uta, utb, x, y, xp, yp)

Returns celestial to terrestrial matrix for given two, 2 part Julian dates (UT1, TT) and CIP coordinates.

Example

julia> celestial_to_terrestrial_matrix(2.4579405e6, 0.0, 2.4579405e6, -0.0007966009351851851, 0.1, 0.3, 0.5, 0.7)
3×3 Rotations.RotMatrix{3,Float64,9}:
  0.279858   0.75876   0.588186
 -0.66989   -0.284531  0.685777
  0.687697  -0.58594   0.428658

AstroBase.EarthAttitude.earth_rotation_angle — Method

earth_rotation_angle(jd1, jd2)

Return Earth rotation angle (radians) for a given UT1 2-part Julian Date (jd1, jd2).

Example

julia> earth_rotation_angle(2.4578265e6, 0.30434616919175345)
4.912208135094597

AstroBase.EarthAttitude.equation_of_equinoxes_00 — Method

equation_of_equinoxes_00(jd1, jd2, epsa, dpsi)

Return equation of equinoxes for given 2 part Julian date (TT), mean obliquity and nutation in longitude.

Example

julia> equation_of_equinoxes_00(2.4578265e6, 0.30440190993249416, 1.5, 1.7)
0.12025324854189404

AstroBase.EarthAttitude.equation_of_equinoxes_94 — Method

equation_of_equinoxes_94(jd1, jd2)

Returns equation of equinoxes for a given 2 part Julian date (TT). (IAU 1994 model)

Example

julia> equation_of_equinoxes_94(2.4578265e6, 0.30434616919175345)
-3.446039061100115e-5

AstroBase.EarthAttitude.equation_of_equinoxes_a00 — Method

equation_of_equinoxes_a00(jd1, jd2)

Returns equation of equinoxes for a given 2 part Julian date(TT).

Example

julia> equation_of_equinoxes_a00(2.4578265e6, 0.30434616919175345)
-3.4482238309105874e-5

AstroBase.EarthAttitude.equation_of_equinoxes_a06 — Method

equation_of_equinoxes_a06(jd1, jd2)

Returns equation of equinoxes for a given 2 part Julian date (TT).

Example

julia> equation_of_equinoxes_a06(2.4578265e6, 0.30434616919175345)
-3.448290610741367e-5

AstroBase.EarthAttitude.equation_of_equinoxes_b00 — Method

equation_of_equinoxes_b00(jd1, jd2)

Returns equation of equinoxes for a given 2 part Julian date(TT).

Example

julia> equation_of_equinoxes_b00(2.4578265e6, 0.30434616919175345)
-3.4482238309105874e-5

AstroBase.EarthAttitude.equation_of_equinoxes_complementary_terms — Method

equation_of_equinoxes_complementary_terms(jd1, jd2)

Returns complementary terms for a given 2 part Julian date (TT).

Example

julia> equationofequinoxescomplementaryterms(2.4578265e6, 0.30434616919175345) 5.706799075288604e-9 ```

AstroBase.EarthAttitude.equation_of_origins — Method

equation_of_origins(rnpb, s)

Returns the equation of origins(radians) for given nutation-bias-precession matrix and the CIO locator.

Example

jldoctest equation_of_origins(rand(3,3), 0.2) 1.7738370040531068

AstroBase.EarthAttitude.fukushima_williams_matrix — Method

fukushima_williams_matrix(gamb, phib, psi, eps)

Returns obliquity of the ecliptic (radians) for a given Julian 2 part date (TT).

Example

julia> fukushima_williams_matrix(0.2,0.3,0.5,0.6)
3×3 RotMatrix{Float64}:
  0.951082   0.21718   0.219716
 -0.274534   0.920305  0.278692
 -0.14168   -0.325378  0.93491

AstroBase.EarthAttitude.general_precession_in_longitude — Method

general_precession_in_longitude(t)

Returns general accumulated precession in longitude for Julian centuries since J2000.0 in TDB.

Example

julia> general_precession_in_longitude(23.0)
0.56362992539

AstroBase.EarthAttitude.greenwich_apparent_sidereal_time06 — Method

greenwich_mean_sidereal_time06(uta, utb, tta, ttb, rnpb)

Returns Greenwich mean sidereal time(radians) for given two, 2 part Julian dates (TT and UT1).

Example

julia> gst06(2.4579405e6, 0.0, 2.4579405e6, -0.0007966009351851851, rand(3,3))
3.738379033673736

AstroBase.EarthAttitude.greenwich_apparent_sidereal_time94 — Method

greenwich_apparent_sidereal_time94(uta, utb)

Returns greenwich appart sidereal time for given 2 part Julian date (UT1).

Example

julia> greenwich_apparent_sidereal_time94(2.4578265e6, 0.30434616919175345)
4.9160197844443445

AstroBase.EarthAttitude.greenwich_apparent_sidereal_time_a00 — Method

greenwich_apparent_sidereal_time_a00(uta, utb, tta, ttb)

Returns Greenwich apparent sidereal time for given two, 2 part Julian date. (UT1, TT)

Example

julia> greenwich_apparent_sidereal_time_a00(2.4579405e6, 0.0, 2.4579405e6, -0.0007966009351851851)
4.959632360790386

AstroBase.EarthAttitude.greenwich_apparent_sidereal_time_a06 — Method

greenwich_apparent_sidereal_time_a06(uta, utb, tta, ttb)

Returns Greenwich apparent sidereal time for given two, 2 part Julian dates. (UT1, TT)

Example

julia> greenwich_apparent_sidereal_time_a06(2.4579405e6, 0.0, 2.4579405e6, -0.0007966009351851851)
4.959632359298287

AstroBase.EarthAttitude.greenwich_apparent_sidereal_time_b00 — Method

greenwich_apparent_sidereal_time_b00(uta, utb, tta, ttb)

Returns Greenwich apparent sidereal time for given two, 2 part Julian date. (UT1, TT)

Example

julia> greenwich_apparent_sidereal_time_b00(2.4579405e6, 0.0, 2.4579405e6, -0.0007966009351851851)
4.95963236232993

AstroBase.EarthAttitude.greenwich_mean_sidereal_time00 — Method

greenwich_mean_sidereal_time00(ut1, ut2, tt1, tt2)

Returns Greenwich mean sidereal time(radians) for given two, 2 part Julian dates (TT and UT1). (consistent with IAU 2000 precession)

Example

julia> greenwich_mean_sidereal_time00(2.4579405e6, 0.0, 2.4579405e6, -0.0007966009351851851)
4.9596733720586075

AstroBase.EarthAttitude.greenwich_mean_sidereal_time06 — Method

greenwich_mean_sidereal_time06(ut1, ut2, tt1, tt2)

Returns Greenwich mean sidereal time(radians) for given two, 2 part Julian dates (TT and UT1). (consistent with IAU 2006 precession)

Example

julia> greenwich_mean_sidereal_time06(2.4579405e6, 0.0, 2.4579405e6, -0.0007966009351851851)
4.959673370568533

AstroBase.EarthAttitude.greenwich_mean_sidereal_time82 — Method

greenwich_mean_sidereal_time82(jd1, jd2)

Returns Greenwich mean sidereal time(radians) for given 2 part Julian dates (UT1). (consistent with IAU 1982 model)

Example

julia> greenwich_mean_sidereal_time82(2.4578265e6, 0.30434616919175345)
4.916054244834956

AstroBase.EarthAttitude.greenwich_mean_sidereal_time_a06 — Method

greenwich_mean_sidereal_time_a06(uta, utb, tta, ttb, rnpb)

Returns greenwich apparent sidereal time for given two, 2 part Julian dates (UT1, TT).

Example

AstroBase.EarthAttitude.mean_elongation — Method

mean_elongation(::Luna, t)

Returns mean elongation of the Moon from the Sun for Julian centuries since J2000.0 in TDB.

Example

julia> mean_elongation(::Luna, 23.0)
2.8012040574296484

AstroBase.EarthAttitude.mean_longitude — Method

mean_longitude(::Earth, t)

Returns mean longitude of Earth for Julian centuries since J2000.0 in TDB.

Example

julia> mean_longitude(earth, 23.0)
1.501718780251826

AstroBase.EarthAttitude.mean_longitude — Method

mean_longitude(::Jupiter, t)

Returns mean longitude of Jupiter for Julian centuries since J2000.0 in TDB.

Example

julia> mean_longitude(jupiter, 23.0)
6.233996285639864

AstroBase.EarthAttitude.mean_longitude — Method

mean_longitude(::Mars, t)

Returns mean longitude of Mars for Julian centuries since J2000.0 in TDB.

Example

julia> mean_longitude(mars, 23.0)
5.276431642365657

AstroBase.EarthAttitude.mean_longitude — Method

mean_longitude(::Mercury, t)

Returns mean longitude of Mercury for Julian centuries since J2000.0 in TDB.

Example

julia> mean_longitude(mercury, 23.0)
2.160150897150834

AstroBase.EarthAttitude.mean_longitude — Method

mean_longitude(::Neptune, t)

Returns mean longitude of Neptune for Julian centuries since J2000.0 in TDB.

Example

julia> mean_longitude(neptune, 23.0)
5.053273953885775

AstroBase.EarthAttitude.mean_longitude — Method

mean_longitude(::Saturn, t)

Returns mean longitude of Saturn for Julian centuries since J2000.0 in TDB.

Example

julia> mean_longitude(saturn, 23.0)
1.3735042049922535

AstroBase.EarthAttitude.mean_longitude — Method

mean_longitude(::Uranus, t)

Returns mean longitude of Uranus for Julian centuries since J2000.0 in TDB.

Example

julia> mean_longitude(uranus, 23.0)
1.5497819750715893

AstroBase.EarthAttitude.mean_longitude — Method

mean_longitude(::Venus, t)

Returns mean longitude of Venus for Julian centuries since J2000.0 in TDB.

Example

julia> mean_longitude(venus, 23.0)
0.9030394378238363

AstroBase.EarthAttitude.mean_longitude_ascending_node — Method

mean_longitude_ascending_node(::Luna, t)

Return fundamental argument for Julian centuries since J2000.0 in TDB.

Example

julia> mean_longitude_ascending_node(luna, 23.0)
4.904897783682109

AstroBase.EarthAttitude.mean_longitude_minus_lan — Method

mean_longitude_minus_lan(::Luna, t)

Returns mean longitude of the Moon minus mean longitude of the ascending node for Julian centuries since J2000.0 in TDB.

Example

julia> mean_longitude_minus_lan(23.0)
3.103138156410118

AstroBase.EarthAttitude.numat — Method

numat(epsa, dpsi, deps)

Returns nutation matrix for a given epsa(mean obliquity), dpsi and deps nutation.

Example

julia> numat(0.7, 1.4, 1.3)
3×3 RotXZX{Float64}(0.7, -1.4, -2.0):
  0.169967  -0.410092   0.896067
 -0.753714   0.531687   0.386296
 -0.634844  -0.741035  -0.218722

AstroBase.EarthAttitude.nutation — Function

nutation(model, ep)

Return the nutation components for a given epoch and model.

Arguments

model: IAU model, one of: iau1980, iau2000a, iau2000b, iau2006

ep: An epoch

Output

δψ: Nutation in longitude (radians)

δϵ: Nutation in obliquity (radians)

Example

julia> ep = TTEpoch(2020, 1, 1)
2020-01-01T00:00:00.000 TT

julia> nutation(iau2006, ep)
(-7.996558232098883e-5, -8.25141288270117e-6)

References

SOFA

AstroBase.EarthAttitude.nutation_matrix80 — Method

nutation_matrix80(jd1, jd2)

Returns nutation matrix for a given 2 part Julian date (TDB).

Example

jldoctest julia> nutation_matrix80(2.4578265e6, 0.30434616919175345) 3×3 Rotations.RotXZX{Float64}(0.409054, 3.75653e-5, -0.409017): 1.0 -3.44666e-5 -1.494e-5 3.44661e-5 1.0 -3.66564e-5 1.49413e-5 3.66559e-5 1.0

AstroBase.EarthAttitude.nutation_matrix_day — Method

nutation_matrix_day(jd1,jd2)

Returns nutation matrix for a given 2 part Julian date(TT).

Example

julia> precession_nutation_b00(2.4578265e6, 0.30434616919175345)[3][5]
3×3 Rotations.RotXZX{Float64}(0.409054, 3.75891e-5, -0.409017):
 1.0         -3.44885e-5  -1.49495e-5
 3.44879e-5   1.0         -3.66577e-5
 1.49508e-5   3.66572e-5   1.0

AstroBase.EarthAttitude.nutation_matrix_day_a00 — Method

nutation_matrix_day_a00(jd1, jd2)

Returns nutation matrix for a given 2 part Julian date(TT).

Example

julia> precession_nutation_a00(2.4578265e6, 0.30434616919175345)[3][5]
3×3 Rotations.RotXZX{Float64}(0.409054, 3.75891e-5, -0.409017):
 1.0         -3.44885e-5  -1.49495e-5
 3.44879e-5   1.0         -3.66577e-5
 1.49508e-5   3.66572e-5   1.0

AstroBase.EarthAttitude.nutation_matrix_day_b00 — Method

nutation_matrix_day_b00(jd1, jd2)

Returns nutation matrix for a given 2 part Julian date(TT).

Example

julia> precession_nutation_b00(2.4578265e6, 0.30434616919175345)[3][5]
3×3 Rotations.RotXZX{Float64}(0.409054, 3.75891e-5, -0.409017):
 1.0         -3.44885e-5  -1.49495e-5
 3.44879e-5   1.0         -3.66577e-5
 1.49508e-5   3.66572e-5   1.0

AstroBase.EarthAttitude.obliquity — Function

obliquity(model, ep)

Return the mean obliquity of the ecliptic for a given epoch and model.

Arguments

model: IAU model, one of: iau1980, iau2006

ep: An epoch

Output

Returns the angle between the ecliptic and mean equator of date in radians.

Example

julia> ep = TTEpoch(2020, 1, 1)
2020-01-01T00:00:00.000 TT

julia> obliquity(iau2006, ep)
0.40904718953841473

References

SOFA

AstroBase.EarthAttitude.polar_motion — Method

polar_motion(rx, ry, sp)

Form the matrix of polar motion for coordinates of the pole (radians).

Example

julia> polar_motion(20, 30, 50)
3×3 RotZYX{Float64}(50.0, -20.0, -30.0):
  0.393785  -0.829946  -0.395124
 -0.10707    0.385514  -0.916469
  0.912945   0.403198   0.0629472

AstroBase.EarthAttitude.precession_fukushima_williams06 — Method

precession_fukushima_williams06(jd1, jd2)

Returns fukushima angles(radians) for a given 2 part Julian date (TT).

Example

julia> precession_fukushima_williams06(2.4578265e6, 0.30434616919175345)
(8.616170933989655e-6, 0.4090536093366178, 0.004201176043952816, 0.409053547482157)

AstroBase.EarthAttitude.precession_nutation00 — Method

precession_nutation00(jd1, jd2, dpsi, deps)

Returns epsa rb rp rbp rn rbpn for a given 2 part Julian date (TT) and nutations.

Example

julia> pn00(2.4578265e6, 0.30434616919175345, 0.2, 0.2)
(0.40905374831590824, [1.0 7.07828e-8 -8.05622e-8; -7.07828e-8 1.0 -3.30604e-8; 8.05622e-8 3.30604e-8 1.0], [0.999991 0.00384587 0.00167106; -0.00384587 0.999993 -3.24146e-6; -0.00167106 -3.18524e-6 0.999999], [0.999991 0.00384594 0.00167098; -0.00384594 0.999993 -3.27421e-6; -0.00167098 -3.1523e-6 0.999999], [0.980067 0.162947 0.113657; -0.182279 0.965066 0.188206; -0.079019 -0.205172 0.975531], [0.979225 0.166314 0.11601; -0.186046 0.964433 0.187765; -0.080656 -0.205447 0.975339])

AstroBase.EarthAttitude.precession_nutation_a00 — Method

precession_nutation_a00(jd1, jd2, dpsi, deps)

Returns dpsi, deps epsa rb rp rbp rn rbpn for a given 2 part Julian date (TT) and nutations.

Example

julia> precession_nutation_a00(2.4578265e6, 0.30434616919175345)
(-3.758986507815969e-5, -3.6659048454874034e-5, (0.40905374831590824, [1.0 7.07828e-8 -8.05622e-8; -7.07828e-8 1.0 -3.30604e-8; 8.05622e-8 3.30604e-8 1.0], [0.999991 0.00384587 0.00167106; -0.00384587 0.999993 -3.24146e-6; -0.00167106 -3.18524e-6 0.999999], [0.999991 0.00384594 0.00167098; -0.00384594 0.999993 -3.27421e-6; -0.00167098 -3.1523e-6 0.999999], [1.0 -3.44892e-5 -1.49498e-5; 3.44886e-5 1.0 -3.66593e-5; 1.4951e-5 3.66588e-5 1.0], [0.999991 0.00381152 0.00165589; -0.00381145 0.999993 -3.98758e-5; -0.00165603 3.35641e-5 0.999999]))

AstroBase.EarthAttitude.precession_nutation_b00 — Method

precession_nutation_b00(jd1, jd2, dpsi, deps)

Returns dpsi, deps epsa rb rp rbp rn rbpn for a given 2 part Julian date (TT) and nutations.

Example

julia> precession_nutation_b00(2.4578265e6, 0.30434616919175345)
(-3.758913944197044e-5, -3.665743003046017e-5, (0.40905374831590824, [1.0 7.07828e-8 -8.05622e-8; -7.07828e-8 1.0 -3.30604e-8; 8.05622e-8 3.30604e-8 1.0], [0.999991 0.00384587 0.00167106; -0.00384587 0.999993 -3.24146e-6; -0.00167106 -3.18524e-6 0.999999], [0.999991 0.00384594 0.00167098; -0.00384594 0.999993 -3.27421e-6; -0.00167098 -3.1523e-6 0.999999], [1.0 -3.44885e-5 -1.49495e-5; 3.44879e-5 1.0 -3.66577e-5; 1.49508e-5 3.66572e-5 1.0], [0.999991 0.00381152 0.00165589; -0.00381146 0.999993 -3.98741e-5; -0.00165603 3.35625e-5 0.999999]))

AstroBase.EarthAttitude.precession_nutation_matrix_a00 — Method

precession_nutation_matrix_a00(jd1, jd2)

Returns classical NBP matrix for a given 2 part Julian date(TT)

Example

julia> precession_nutation_matrix_a00(2.4578265e6, 0.30434616919175345)
3×3 Rotations.RotMatrix{3,Float64,9}:
  0.999991    0.00381152   0.00165589
 -0.00381145  0.999993    -3.98758e-5
 -0.00165603  3.35641e-5   0.999999

AstroBase.EarthAttitude.precession_nutation_matrix_a06 — Method

precession_nutation_matrix_a06(jd1, jd2)

Returns classical NBP matrix for a given 2 part Julian date(TT)

Example

julia> precession_nutation_matrix_a06(2.4578265e6, 0.30434616919175345)
3×3 Rotations.RotMatrix{3,Float64,9}:
  0.999991    0.00381151   0.00165589
 -0.00381145  0.999993    -3.98694e-5
 -0.00165602  3.35578e-5   0.999999

AstroBase.EarthAttitude.precession_nutation_matrix_b00 — Method

precession_nutation_matrix_b00(jd1, jd2)

Returns classical NBP matrix for a given 2 part Julian date(TT)

Example

julia> precession_nutation_matrix_b00(2.4578265e6, 0.30434616919175345)
3×3 Rotations.RotMatrix{3,Float64,9}:
  0.999991    0.00381152   0.00165589
 -0.00381146  0.999993    -3.98741e-5
 -0.00165603  3.35625e-5   0.999999

AstroBase.EarthAttitude.precession_rate_part_of_nutation — Method

precession_rate_part_of_nutation(jd1, jd2)

Returns precession corrections for a given 2 part Julian date (TT).

Example

julia> precession_rate_part_of_nutation(2400000.5, 53736)
-0.8716465172668347629e-7, -0.7342018386722813087e-8

AstroBase.EarthAttitude.s00 — Function

s00(jd1, jd2, x, y)

Returns Celestial Intermediate Origin(CIO) for a given 2-part Julian date (jd1, jd2) with CIP coordinates (x, y). Compatible with IAU-2000 precession-nutation.

Example

julia> AstroBase.s00(2.4578265e6, 0.30434616919175345, 20, 50)
-500.00000000383193

AstroBase.EarthAttitude.s00a — Method

s00a(jd1, jd2)

Returns the CIO locator (radians) for a given 2 part Julian date (TT).

Example

julia> s00a(2.4578265e6, 0.30434616919175345)
2.9183401976144617e-8

AstroBase.EarthAttitude.s00b — Method

s00b(jd1, jd2)

Returns the CIO locator (radians) for a given 2 part Julian date (TT).

Example

julia> s00b(2.4578265e6, 0.30434616919175345)
2.9182068810695328e-8

AstroBase.EarthAttitude.s06 — Function

s06(jd1, jd2, x, y)

Returns Celestial Intermediate Origin(CIO) for a given 2-part Julian date (jd1, jd2) with CIP coordinates (x, y)

Example

julia> AstroBase.s06(2.4578265e6, 0.30434616919175345, 20, 50) -500.00000000383193 ```

AstroBase.EarthAttitude.s06a — Method

s06a(jd1, jd2)

Returns the CIO locator for a given 2 part Julian date (TT).

Example

julia> s06a(2.4578265e6, 0.30434616919175345)
2.917767075719633e-8

AstroBase.EarthAttitude.tio_locator — Method

tio_locator(jd1, jd2)

Returns TIO locator s' position for a given TT 2-part Julian date (jd1, jd2).

Example

julia> AstroBase.tio_locator(2.4578265e6, 0.30434616919175345)
-3.9189245827947945e-11

AstroBase.EarthAttitude.xy06 — Method

xy06(jd1, jd2)

Returns X, Y coordinates of celestial intermediate pole for a given 2-part Julian date in TT.

Example

julia> xy06(2.4578265e6, 0.30440190993249416)
(0.0016558850230577835, -3.986943362456243e-5)

AstroBase.EarthAttitude.xys00a — Method

xys00a(jd1, jd2)

Returns Celestial intermediate pole (x, y) and the CIO locator for a given 2 part Julian date. (TT)

Example

julia> xys00a(2.4578265e6, 0.30434616919175345)
(0.001655885227051451, -3.987617878202287e-5, 2.9183401976144617e-8)

AstroBase.EarthAttitude.xys00b — Method

xys00a(jd1, jd2)

Returns Celestial intermediate pole (x, y) and the CIO locator for a given 2 part Julian date. (TT)

Example

julia> xys00b(2.4578265e6, 0.30434616919175345)
(0.001655885521808701, -3.987456146931851e-5, 2.9182068810695328e-8)