Earth Attitude

AstroBase.EarthAttitude.iau1980 — Constant

`iau1980`

The singleton instance of type IAU1980, representing the IAU 1980 family of models.

AstroBase.EarthAttitude.iau1982 — Constant

`iau1982`

The singleton instance of type IAU1982, representing the IAU 1982 family of models.

AstroBase.EarthAttitude.iau1994 — Constant

`iau1994`

The singleton instance of type IAU1994, representing the IAU 1994 family of models.

AstroBase.EarthAttitude.iau2000 — Constant

`iau2000`

The singleton instance of type IAU2000, representing the IAU 2000 family of models.

AstroBase.EarthAttitude.iau2000a — Constant

`iau2000a`

The singleton instance of type IAU2000A, representing the IAU 2000A family of models.

AstroBase.EarthAttitude.iau2000b — Constant

`iau2000b`

The singleton instance of type IAU2000B, representing the IAU 2000B family of models.

AstroBase.EarthAttitude.iau2006 — Constant

`iau2006`

The singleton instance of type IAU2006, representing the IAU 2006 family of models.

AstroBase.EarthAttitude.iau2006a — Constant

`iau2006a`

The singleton instance of type IAU2006a, representing the IAU 2006A family of models.

AstroBase.EarthAttitude.cip_coords — Method

cip_coords(rbpn)

Extract from the bias-precession-nutation matrix the X,Y coordinates of the Celestial Intermediate Pole.

References

ERFA

AstroBase.EarthAttitude.nutation — Function

nutation(model, ep)

Return the nutation components for a given epoch and model.

Arguments

model: IAU model, one of: iau1980, iau2000a, iau2000b, iau2006

ep: An epoch

Output

δψ: Nutation in longitude (radians)

δϵ: Nutation in obliquity (radians)

Example

julia> ep = TTEpoch(2020, 1, 1)
2020-01-01T00:00:00.000 TT

julia> nutation(iau2006, ep)
(-7.996558232098883e-5, -8.25141288270117e-6)

References

SOFA

AstroBase.EarthAttitude.obliquity — Function

obliquity(model, ep)

Return the mean obliquity of the ecliptic for a given epoch and model.

Arguments

model: IAU model, one of: iau1980, iau2006

ep: An epoch

Output

Returns the angle between the ecliptic and mean equator of date in radians.

Example

julia> ep = TTEpoch(2020, 1, 1)
2020-01-01T00:00:00.000 TT

julia> obliquity(iau2006, ep)
0.40904718953841473

References

SOFA